requirements.txt
django==1.8.4
django-registration-redux>=1.2
